updated bsp

+++ cpp02/ex03/src/bsp.cpp
@@ -6,7 +6,7 @@
 /*   By: jpelline <jpelline@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/21 12:45:06 by jpelline          #+#    #+#             */
-/*   Updated: 2025/08/21 12:45:45 by jpelline         ###   ########.fr       */
+/*   Updated: 2025/09/04 13:56:29 by jpelline         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,11 +16,11 @@ static Fixed triangle_area(const Fixed x1, const Fixed y1,
 				  const Fixed x2, const Fixed y2,
 				  const Fixed x3, const Fixed y3) {
 
-	Fixed x1Part = x1.toFloat() * (y2.toFloat() - y3.toFloat());
-	Fixed x2Part = x2.toFloat() * (y3.toFloat() - y1.toFloat());
-	Fixed x3Part = x3.toFloat() * (y1.toFloat() - y2.toFloat());
+	Fixed x1Part = x1 * (y2 - y3);
+	Fixed x2Part = x2 * (y3 - y1);
+	Fixed x3Part = x3 * (y1 - y2);
 
-	Fixed area = (x1Part.toFloat() + x2Part.toFloat() + x3Part.toFloat()) / 2.0f;
+	Fixed area = (x1Part + x2Part + x3Part) / 2.0f;
 
 	if (area < 0)
 		area = -area;
